Update main to production

+++ camastersoftware.com/app/Views/firm_panel/mainpage/getDateActs.php
@@ -178,7 +178,7 @@
                                                                 $datediff = $now - $your_date;
                                                                 
                                                                 if($your_date>$now)
-                                                                    echo abs(round($datediff / (60 * 60 * 24)));
+                                                                    echo abs((float)round($datediff / (60 * 60 * 24)));
                                                                 else
                                                                     echo "-";
                                                             ?>
