+++ camastersoftware.com/app/Views/super_admin/main_menu/getDateActs.php
@@ -199,7 +199,7 @@
                                                                 $datediff = $now - $your_date;
                                                                 
                                                                 if($your_date>$now)
-                                                                    echo abs(round($datediff / (60 * 60 * 24)));
+                                                                    echo abs((float)round($datediff / (60 * 60 * 24)));
                                                                 else
                                                                     echo "-";
                                                             ?>
